% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_dt.R
\name{print_dt}
\alias{print_dt}
\title{Print a Downloadable Table}
\usage{
print_dt(x)
}
\description{
Print a data table in a R Markdown document as described in \url{https://martinctc.github.io/blog/vignette-downloadable-tables-in-rmarkdown-with-the-dt-package/}
}
